+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.8.1
 
  - The blob spell will now auto-undo after 30 seconds.
+ - Some fixes to the Essentials integration.
 
 ## 2.8.0